+++ pkg/hexapolicy/conditions/parser/parser.go
